added graphs.cpp & formatted

+++ modularMultiplication.cpp
@@ -1,17 +1,16 @@
-ll const MOD = 1e9 + 7;
+long long const MOD = 1e9 + 7;
 
-
-ll mulMOD(ll a, ll b) 
-{ 
-    ll mod = MOD;
-    ll res = 0;
-    a = a % mod; 
-    while (b > 0) 
-    { 
-        if (b % 2 == 1) 
-            res = (res + a) % mod; 
-        a = (a * 2) % mod; 
-        b /= 2; 
-    } 
-    return res % mod; 
-} 
+long long mulMOD(long long a, long long b)
+{
+    long long mod = MOD;
+    long long res = 0;
+    a = a % mod;
+    while (b > 0)
+    {
+        if (b % 2 == 1)
+            res = (res + a) % mod;
+        a = (a * 2) % mod;
+        b /= 2;
+    }
+    return res % mod;
+}
